introduce indexing features
Introduced features for indexing Java files in a project to offer better
import suggestions. The index is an import tree saved in memory that is
populated from a variety of sources. The index can be populated from a
specific buffer, from project files (recursive directory traversal), or
can be recovered from a file (configurable).